requirements.txt
flask
flask_sqlalchemy
SQLAlchemy-migrate